fix: made fixations to make response for main found board to work
