+++ src/main/java/com/animalfarm/animalfarm_back/controller/request/BoardAddRequest.java
@@ -16,7 +16,6 @@ public class BoardAddRequest {
     private String detailLocation;
     private String phoneNum;
     private String content;
-    private int boardType;
 
     public BoardAddRequest(String content, String phoneNum, String detailLocation, double latitude, double longitude, String location, int category, String title) {
         this.content = content;
